+++ SBLib/Classes/Model/Database/Queue/DataAppDB+Queue.m
@@ -174,7 +174,7 @@
 }
 
 //设置某条整型数据
-- (void)setIntValue:(nullable NSString *)dataType dataKey:(nullable NSString *)dataKey dataValue:(int)dataValue withBlock:(nullable void(^)(FMDatabase *__nonnull, sqlite3_int64 retVal))completeBlock {
+- (void)setIntValue:(nullable NSString *)dataType dataKey:(nullable NSString *)dataKey dataValue:(NSInteger)dataValue withBlock:(nullable void(^)(FMDatabase *__nonnull, sqlite3_int64 retVal))completeBlock {
     [self setItemValue:_dbIntValueTable dataType:dataType dataKey:dataKey dataValue:@(dataValue) withBlock:completeBlock];
 }
